Implements allow, deny and border blocks
