+++ src/main/java/cn/nukkit/block/BlockAllow.java
@@ -1,0 +1,69 @@
+package cn.nukkit.block;
+
+import cn.nukkit.Player;
+import cn.nukkit.item.Item;
+import cn.nukkit.math.BlockFace;
+import cn.nukkit.math.Vector3;
+
+public class BlockAllow extends BlockSolid {
+    @Override
+    public int getId() {
+        return ALLOW;
+    }
+
+    @Override
+    public double getHardness() {
+        return -1;
+    }
+
+    @Override
+    public double getResistance() {
+        return 18000000;
+    }
+
+    @Override
+    public String getName() {
+        return "Allow";
+    }
+
+    @Override
+    public boolean isBreakable(Item item) {
+        return false;
+    }
+
+    @Override
+    public boolean canBePushed() {
+        return false;
+    }
+
+    @Override
+    public boolean canBePulled() {
+        return false;
+    }
+
+    @Override
+    public boolean canHarvestWithHand() {
+        return false;
+    }
+
+    @Override
+    public boolean place(Item item, Block block, Block target, BlockFace face, double fx, double fy, double fz, Player player) {
+        if (!player.isOp()) {
+            return false;
+        }
+        return super.place(item, block, target, face, fx, fy, fz, player);
+    }
+
+    @Override
+    public boolean isBreakable(Vector3 vector, int layer, BlockFace face, Item item, Player player, boolean setBlockDestroy) {
+        if (player != null && !player.isOp()) {
+            return false;
+        }
+        return super.isBreakable(vector, layer, face, item, player, setBlockDestroy);
+    }
+
+    @Override
+    public Item[] getDrops(Item item) {
+        return new Item[0];
+    }
+}
